requirements.txt
backoff>=2.2.0
databricks-connect~=16.1.1
databricks-agents>=0.15.0
databricks-langchain
databricks-sdk
databricks-vectorsearch>=0.52
docling>=2.15.0
jupyter>=1.1.1
langchain-community>=0.3.14
langchain-core>=0.3.29
langchain-openai>=0.3
langgraph>=0.2.63
mlflow>=2.4.0
nltk>=3.9.1
numpy>=1.26.0
openai>=1.35.0
pydantic>=2.1.1
pyyaml>=6.0.1
transformers>=4.42
streamlit>=1.25.0